replacing long type with double in tables combobox as it was meant to be. The long was already parsed as a double.

+++ src/logdruid/ui/table/EventRecordingEditorTable.java
@@ -232,7 +232,7 @@ public class EventRecordingEditorTable extends JPanel {
 		comboBox.addItem("word");
 		comboBox.addItem("string");
 		comboBox.addItem("integer");
-		comboBox.addItem("long");
+		comboBox.addItem("double");
 		comboBox.addItem("date");
 		comboBox.addItem("percent");
