+++ src/logdruid/ui/table/MetadataRecordingEditorTable.java
@@ -207,7 +207,7 @@ public class MetadataRecordingEditorTable extends JPanel {
 		comboBox.addItem("word");
 		comboBox.addItem("string");
 		comboBox.addItem("integer");
-		comboBox.addItem("long");
+		comboBox.addItem("double");
 		comboBox.addItem("date");
 		comboBox.addItem("percent");
 		TypeColumn.setCellEditor(new DefaultCellEditor(comboBox));
