+++ src/logdruid/ui/table/StatRecordingEditorTable.java
@@ -211,7 +211,7 @@ public class StatRecordingEditorTable extends JPanel {
 		comboBox.addItem("word");
 		comboBox.addItem("string");
 		comboBox.addItem("integer");
-		comboBox.addItem("long");
+		comboBox.addItem("double");
 		comboBox.addItem("date");
 		comboBox.addItem("percent");
 		TypeColumn.setCellEditor(new DefaultCellEditor(comboBox));
